move tilelink-agnostic counters from uncore to util package
